refactoring: main app 컴포넌트 관련 파일을 /src/App/.. 으로 이동
