+++ frontend/main-app/src/index.js
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import "./index.css";
-import App from "./App";
+import App from "./App/App.js";
 import * as serviceWorker from "./libs/serviceWorker.js";
 
 ReactDOM.render(<App />, document.getElementById("root"));
